separated the machine learning code into a separate app
