+++ machine_learning/.gitignore
@@ -1,0 +1,2 @@
+__pycache__
+management\commands\__pycache__
