+++ machine_learning/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class MachinelearningConfig(AppConfig):
+    name = 'machine_learning'
