+++ machine_learning/urls.py
@@ -1,0 +1,12 @@
+from django.contrib import admin
+from django.urls import path
+from django.conf.urls import url
+
+
+app_name = 'machine_learning'
+
+urlpatterns = [
+
+
+]
+
